+++ benches/mul.rs
@@ -1,6 +1,6 @@
 use std::fmt::Debug;
 
-use const_decimal::{Decimal, Integer};
+use const_decimal::{Decimal, ScaledInteger};
 use criterion::measurement::WallTime;
 use criterion::{black_box, BatchSize, BenchmarkGroup};
 use num_traits::PrimInt;
@@ -13,7 +13,7 @@ pub fn bench_all<const D: u8, I>(
     lo_strategy: impl Strategy<Value = I> + Clone,
     hi_strategy: impl Strategy<Value = I> + Clone,
 ) where
-    I: Integer<D> + Debug,
+    I: ScaledInteger<D> + Debug,
 {
     primitive_mul::<I>(group, lo_strategy.clone(), "lo");
     decimal_mul::<D, I>(group, lo_strategy, "lo");
@@ -46,7 +46,7 @@ fn decimal_mul<const D: u8, I>(
     strategy: impl Strategy<Value = I> + Clone,
     strategy_label: &str,
 ) where
-    I: Integer<D> + Debug,
+    I: ScaledInteger<D> + Debug,
 {
     // Use proptest to generate arbitrary input values.
     let mut runner = TestRunner::deterministic();