+++ src/decimal.rs
@@ -1,7 +1,7 @@
 use std::cmp::Ordering;
 use std::ops::{Add, AddAssign, Div, DivAssign, Mul, MulAssign, Neg, Sub, SubAssign};
 
-use crate::integer::{Integer, SignedInteger};
+use crate::integer::{ScaledInteger, SignedScaledInteger};
 
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 #[cfg_attr(feature = "borsh", derive(borsh::BorshSerialize, borsh::BorshDeserialize))]
@@ -11,7 +11,7 @@ pub struct Decimal<I, const D: u8>(pub I);
 
 impl<I, const D: u8> Decimal<I, D>
 where
-    I: Integer<D>,
+    I: ScaledInteger<D>,
 {
     pub const ZERO: Decimal<I, D> = Decimal(I::ZERO);
     pub const ONE: Decimal<I, D> = Decimal(I::SCALING_FACTOR);
@@ -31,10 +31,52 @@ where
         Decimal(I::max_value())
     }
 
+    // TODO: Add `try_from_scaled() -> Result<>`.
+
+    /// Losslessly converts a scaled integer to this type.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the integer cannot be represented without precision
+    /// loss/overflow.
+    ///
+    /// # Examples
+    ///
+    /// ```rust
+    /// use const_decimal::Decimal;
+    ///
+    /// let five = Decimal::<u64, 3>::from_scaled(5, 0);
+    /// assert_eq!(five, Decimal::TWO + Decimal::TWO + Decimal::ONE);
+    /// assert_eq!(five.0, 5000);
+    /// ```
     pub fn from_scaled(integer: I, scale: u8) -> Self {
         match scale.cmp(&D) {
-            Ordering::Greater => todo!(),
-            Ordering::Less => todo!(),
+            Ordering::Greater => {
+                // SAFETY: We know `scale > D` so this cannot underflow.
+                #[allow(clippy::arithmetic_side_effects)]
+                let divisor = I::TEN.pow(u32::from(scale - D));
+
+                // SAFETY: `divisor` cannot be zero as `x.pow(y)` cannot return 0.
+                #[allow(clippy::arithmetic_side_effects)]
+                let remainder = integer % divisor;
+                assert!(
+                    remainder == I::ZERO,
+                    "Cast would lose precision; input={integer}; scale={scale}; divisor={divisor}"
+                );
+
+                // SAFETY: `Decimal::mul` panics on divide by zero.
+                #[allow(clippy::arithmetic_side_effects)]
+                Decimal(integer / divisor)
+            }
+            Ordering::Less => {
+                // SAFETY: We know `scale < D` so this cannot underflow.
+                #[allow(clippy::arithmetic_side_effects)]
+                let multiplier = I::TEN.pow(u32::from(D - scale));
+
+                // SAFETY: `Decimal::mul` panics on overflow.
+                #[allow(clippy::arithmetic_side_effects)]
+                Decimal(integer * multiplier)
+            }
             Ordering::Equal => Decimal(integer),
         }
     }
@@ -46,7 +88,7 @@ where
 
 impl<I, const D: u8> Add for Decimal<I, D>
 where
-    I: Integer<D>,
+    I: ScaledInteger<D>,
 {
     type Output = Self;
 
@@ -58,7 +100,7 @@ where
 
 impl<I, const D: u8> Sub for Decimal<I, D>
 where
-    I: Integer<D>,
+    I: ScaledInteger<D>,
 {
     type Output = Self;
 
@@ -70,7 +112,7 @@ where
 
 impl<I, const D: u8> Mul for Decimal<I, D>
 where
-    I: Integer<D>,
+    I: ScaledInteger<D>,
 {
     type Output = Self;
 
@@ -82,7 +124,7 @@ where
 
 impl<I, const D: u8> Div for Decimal<I, D>
 where
-    I: Integer<D>,
+    I: ScaledInteger<D>,
 {
     type Output = Self;
 
@@ -94,7 +136,7 @@ where
 
 impl<I, const D: u8> Neg for Decimal<I, D>
 where
-    I: SignedInteger<D>,
+    I: SignedScaledInteger<D>,
 {
     type Output = Self;
 
@@ -105,7 +147,7 @@ where
 
 impl<I, const D: u8> AddAssign for Decimal<I, D>
 where
-    I: Integer<D>,
+    I: ScaledInteger<D>,
 {
     #[inline]
     fn add_assign(&mut self, rhs: Self) {
@@ -115,7 +157,7 @@ where
 
 impl<I, const D: u8> SubAssign for Decimal<I, D>
 where
-    I: Integer<D>,
+    I: ScaledInteger<D>,
 {
     #[inline]
     fn sub_assign(&mut self, rhs: Self) {
@@ -125,7 +167,7 @@ where
 
 impl<I, const D: u8> MulAssign for Decimal<I, D>
 where
-    I: Integer<D>,
+    I: ScaledInteger<D>,
 {
     #[inline]
     fn mul_assign(&mut self, rhs: Self) {
@@ -135,7 +177,7 @@ where
 
 impl<I, const D: u8> DivAssign for Decimal<I, D>
 where
-    I: Integer<D>,
+    I: ScaledInteger<D>,
 {
     #[inline]
     fn div_assign(&mut self, rhs: Self) {
@@ -145,10 +187,11 @@ where
 
 #[cfg(test)]
 mod tests {
+    use std::fmt::Debug;
     use std::ops::Shr;
 
     use malachite::num::basic::traits::Zero;
-    use malachite::Rational;
+    use malachite::{Integer, Rational};
     use paste::paste;
     use proptest::prelude::*;
 
@@ -402,13 +445,18 @@ mod tests {
                 fn [<differential_fuzz_ $underlying _ $decimals _div_assign>]() {
                     differential_fuzz_div_assign::<$underlying, $decimals>();
                 }
+
+                #[test]
+                fn [<differential_fuzz_ $underlying _ $decimals _from_scaled>]() {
+                    differential_fuzz_from_scaled::<$underlying, $decimals>();
+                }
             }
         };
     }
 
     fn differential_fuzz_add<I, const D: u8>()
     where
-        I: Integer<D> + Arbitrary + std::panic::RefUnwindSafe,
+        I: ScaledInteger<D> + Arbitrary + std::panic::RefUnwindSafe,
         Rational: From<Decimal<I, D>>,
     {
         proptest!(|(a: Decimal<I, D>, b: Decimal<I, D>)| {
@@ -424,7 +472,7 @@ mod tests {
 
     fn differential_fuzz_sub<I, const D: u8>()
     where
-        I: Integer<D> + Arbitrary + std::panic::RefUnwindSafe,
+        I: ScaledInteger<D> + Arbitrary + std::panic::RefUnwindSafe,
         Rational: From<Decimal<I, D>>,
     {
         proptest!(|(a: Decimal<I, D>, b: Decimal<I, D>)| {
@@ -440,7 +488,7 @@ mod tests {
 
     fn differential_fuzz_mul<I, const D: u8>()
     where
-        I: Integer<D> + Arbitrary + std::panic::RefUnwindSafe + Into<malachite::Integer>,
+        I: ScaledInteger<D> + Arbitrary + std::panic::RefUnwindSafe + Into<Integer>,
         Rational: From<Decimal<I, D>>,
     {
         proptest!(|(a: Decimal<I, D>, b: Decimal<I, D>)| {
@@ -451,9 +499,9 @@ mod tests {
             let reference_out = Rational::from(a) * Rational::from(b);
 
             // If the multiplication contains truncation ignore it.
-            let scaling: malachite::Integer = Decimal::<I, D>::SCALING_FACTOR.into();
-            let divisor = malachite::Integer::from(reference_out.denominator_ref());
-            if scaling % divisor != malachite::Integer::ZERO {
+            let scaling: Integer = Decimal::<I, D>::SCALING_FACTOR.into();
+            let divisor = Integer::from(reference_out.denominator_ref());
+            if scaling % divisor != Integer::ZERO {
                 // TODO: Can we assert they are within N of each other?
                 return Ok(());
             }
@@ -464,7 +512,7 @@ mod tests {
 
     fn differential_fuzz_div<I, const D: u8>()
     where
-        I: Integer<D> + Arbitrary + std::panic::RefUnwindSafe + Into<malachite::Integer>,
+        I: ScaledInteger<D> + Arbitrary + std::panic::RefUnwindSafe + Into<Integer>,
         Rational: From<Decimal<I, D>>,
     {
         proptest!(|(a: Decimal<I, D>, b: Decimal<I, D>)| {
@@ -479,9 +527,9 @@ mod tests {
             let reference_out = Rational::from(a) / Rational::from(b);
 
             // If the division contains truncation ignore it.
-            let scaling: malachite::Integer = Decimal::<I, D>::SCALING_FACTOR.into();
-            let divisor = malachite::Integer::from(reference_out.denominator_ref());
-            if scaling % divisor != malachite::Integer::ZERO {
+            let scaling: Integer = Decimal::<I, D>::SCALING_FACTOR.into();
+            let divisor = Integer::from(reference_out.denominator_ref());
+            if scaling % divisor != Integer::ZERO {
                 // TODO: Can we assert they are within N of each other?
                 return Ok(());
             }
@@ -492,7 +540,7 @@ mod tests {
 
     fn differential_fuzz_add_assign<I, const D: u8>()
     where
-        I: Integer<D> + Arbitrary + std::panic::RefUnwindSafe,
+        I: ScaledInteger<D> + Arbitrary + std::panic::RefUnwindSafe,
         Rational: From<Decimal<I, D>>,
     {
         proptest!(|(a: Decimal<I, D>, b: Decimal<I, D>)| {
@@ -513,7 +561,7 @@ mod tests {
 
     fn differential_fuzz_sub_assign<I, const D: u8>()
     where
-        I: Integer<D> + Arbitrary + std::panic::RefUnwindSafe,
+        I: ScaledInteger<D> + Arbitrary + std::panic::RefUnwindSafe,
         Rational: From<Decimal<I, D>>,
     {
         proptest!(|(a: Decimal<I, D>, b: Decimal<I, D>)| {
@@ -534,7 +582,7 @@ mod tests {
 
     fn differential_fuzz_mul_assign<I, const D: u8>()
     where
-        I: Integer<D> + Arbitrary + std::panic::RefUnwindSafe + Into<malachite::Integer>,
+        I: ScaledInteger<D> + Arbitrary + std::panic::RefUnwindSafe + Into<Integer>,
         Rational: From<Decimal<I, D>>,
     {
         proptest!(|(a: Decimal<I, D>, b: Decimal<I, D>)| {
@@ -550,9 +598,9 @@ mod tests {
             let reference_out = Rational::from(a) * Rational::from(b);
 
             // If the multiplication contains truncation ignore it.
-            let scaling: malachite::Integer = Decimal::<I, D>::SCALING_FACTOR.into();
-            let divisor = malachite::Integer::from(reference_out.denominator_ref());
-            if scaling % divisor != malachite::Integer::ZERO {
+            let scaling: Integer = Decimal::<I, D>::SCALING_FACTOR.into();
+            let divisor = Integer::from(reference_out.denominator_ref());
+            if scaling % divisor != Integer::ZERO {
                 // TODO: Can we assert they are within N of each other?
                 return Ok(());
             }
@@ -563,7 +611,7 @@ mod tests {
 
     fn differential_fuzz_div_assign<I, const D: u8>()
     where
-        I: Integer<D> + Arbitrary + std::panic::RefUnwindSafe + Into<malachite::Integer>,
+        I: ScaledInteger<D> + Arbitrary + std::panic::RefUnwindSafe + Into<Integer>,
         Rational: From<Decimal<I, D>>,
     {
         proptest!(|(a: Decimal<I, D>, b: Decimal<I, D>)| {
@@ -579,14 +627,49 @@ mod tests {
             let reference_out = Rational::from(a) / Rational::from(b);
 
             // If the division contains truncation ignore it.
-            let scaling: malachite::Integer = Decimal::<I, D>::SCALING_FACTOR.into();
-            let divisor = malachite::Integer::from(reference_out.denominator_ref());
-            if scaling % divisor != malachite::Integer::ZERO {
+            let scaling: Integer = Decimal::<I, D>::SCALING_FACTOR.into();
+            let divisor = Integer::from(reference_out.denominator_ref());
+            if scaling % divisor != Integer::ZERO {
                 // TODO: Can we assert they are within N of each other?
                 return Ok(());
             }
 
             assert_eq!(Rational::from(out), reference_out);
+        });
+    }
+
+    fn differential_fuzz_from_scaled<I, const D: u8>()
+    where
+        I: ScaledInteger<D> + Arbitrary + std::panic::RefUnwindSafe + Into<Integer> + TryInto<u64>,
+        Rational: From<I> + From<Decimal<I, D>>,
+        <I as TryInto<u64>>::Error: Debug,
+    {
+        proptest!(|(integer: I, decimals_percent in 0..100u64)| {
+            let max_decimals: u64 = crate::algorithms::log10(I::max_value()).try_into().unwrap();
+            let decimals = u8::try_from(decimals_percent * max_decimals / 100).unwrap();
+            let scaling = I::TEN.pow(decimals as u32);
+
+            let out = std::panic::catch_unwind(|| {
+                Decimal::from_scaled(integer, decimals)
+            });
+            let reference_out = Rational::from_integers(integer.into(), scaling.into());
+
+            match out {
+                Ok(out) => assert_eq!(Rational::from(out), reference_out),
+                Err(_) => {
+                    let scaling: Integer = Decimal::<I, D>::SCALING_FACTOR.into();
+                    let remainder = &scaling % Integer::from(reference_out.denominator_ref());
+                    let information = &reference_out * Rational::from(scaling);
+
+                    assert!(
+                        remainder != 0
+                            || information > Rational::from(I::max_value())
+                            || information < Rational::from(I::min_value()) ,
+                        "Failed to parse valid input; integer={integer}; input_scale={decimals}; \
+                        output_scale={D}",
+                    );
+                }
+            }
         });
     }