+++ src/integer.rs
@@ -1,6 +1,6 @@
 use std::fmt::Display;
 use std::num::ParseIntError;
-use std::ops::{Not, Shr};
+use std::ops::{AddAssign, DivAssign, Not, Shr};
 use std::str::FromStr;
 
 use num_traits::{CheckedNeg, CheckedRem, ConstOne, ConstZero, One, PrimInt, WrappingAdd};
@@ -8,7 +8,7 @@ use num_traits::{CheckedNeg, CheckedRem, ConstOne, ConstZero, One, PrimInt, Wrap
 use crate::cheats::Cheats;
 use crate::full_mul_div::FullMulDiv;
 
-pub trait Integer<const D: u8>:
+pub trait ScaledInteger<const D: u8>:
     PrimInt
     + ConstZero
     + ConstOne
@@ -17,6 +17,8 @@ pub trait Integer<const D: u8>:
     + CheckedRem<Output = Self>
     + Not<Output = Self>
     + Shr<u32, Output = Self>
+    + AddAssign
+    + DivAssign
     + Display
     + FromStr<Err = ParseIntError>
     + Cheats<D>
@@ -24,7 +26,7 @@ pub trait Integer<const D: u8>:
 {
 }
 
-impl<I, const D: u8> Integer<D> for I where
+impl<I, const D: u8> ScaledInteger<D> for I where
     I: PrimInt
         + ConstZero
         + ConstOne
@@ -33,6 +35,8 @@ impl<I, const D: u8> Integer<D> for I where
         + CheckedRem<Output = Self>
         + Not<Output = Self>
         + Shr<u32, Output = Self>
+        + AddAssign
+        + DivAssign
         + Display
         + FromStr<Err = ParseIntError>
         + Cheats<D>
@@ -40,6 +44,6 @@ impl<I, const D: u8> Integer<D> for I where
 {
 }
 
-pub trait SignedInteger<const D: u8>: Integer<D> + CheckedNeg {}
+pub trait SignedScaledInteger<const D: u8>: ScaledInteger<D> + CheckedNeg {}
 
-impl<I, const D: u8> SignedInteger<D> for I where I: Integer<D> + CheckedNeg {}
+impl<I, const D: u8> SignedScaledInteger<D> for I where I: ScaledInteger<D> + CheckedNeg {}