+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to `const-decimal`.
 
 - Added `BorshSerialize` and `BorshDeserialize` behind `borsh` feature flag.
 - Added `AddAssign`, `SubAssign`, `MulAssign`, and `DivAssign`.
+- Implemented `Decimal::from_scaled(integer: I, decimals: u8)`.
 
 ## 0.1.0